requirements_test.txt
# Strictly for tests
-r requirements_dev.txt
aiodiscover
ulid_transform
pytest
coverage
pytest-asyncio
pytest-homeassistant-custom-component